Crate ps_sig

An implementation of the Pointcheval-Sanders signature scheme https://eprint.iacr.org/2015/525

Structs

BlindedSignature
Type wrapper around a signature, indicating that it is a blinded variant.

BlindingRandomness
Randomness used to blind a signature.

KnownMessage
A message to sign. The PS scheme allows signing both a known message, where the message is a vector of values to be signed, and also an unknown message, which is a single value constructed in a special way.

PublicKey
PS public key. The documentation of the fields assumes the secret key is $(x, y_1, …, y_n)$ (see specification).

SecretKey
A secret key

SigRetrievalRandomness
Randomness used to retrieve signature on the message from signature on an unknown message.

Signature
A signature on a KnownMessage.

UnknownMessage
A message to sign. In contrast to KnownMessage this is a single group element that must be constructed in a special way. The idea is that this message is a commitment to some values. The person signing only knows the commitment, but does not know the values they are signing directly.